# artifacts/api-server/src/routes/products.ts
import { Router, type IRouter } from "express";
import { prisma, toNumber, type Prisma } from "@workspace/db";
import {
  CreateProductBody,
  UpdateProductBody,
  UpdateProductParams,
  GetProductParams,
  GetProductResponse,
  UpdateProductResponse,
  GetFeaturedProductsResponse,
  ListProductsQueryParams,
  DeleteProductParams,
} from "@workspace/api-zod";
import { requireClerkAuth } from "../lib/auth";
import { canAccessSupplier, getAuthenticatedDbUser, isAdmin, isSellerOrAdmin } from "../lib/authorize";

const router: IRouter = Router();

type ProductRow = Prisma.ProductGetPayload<object>;

async function enrichProduct(product: ProductRow) {
  const [supplier, category] = await Promise.all([
    prisma.supplier.findUnique({
      where: { id: product.supplierId },
      select: { companyName: true, verified: true, location: true },
    }),
    prisma.category.findUnique({
      where: { id: product.categoryId },
      select: { name: true },
    }),
  ]);

  return {
    ...product,
    minPrice: toNumber(product.minPrice) ?? 0,
    maxPrice: toNumber(product.maxPrice) ?? 0,
    rating: toNumber(product.rating),
    supplierName: supplier?.companyName ?? null,
    supplierVerified: supplier?.verified ?? null,
    supplierLocation: supplier?.location ?? null,
    categoryName: category?.name ?? null,
    images: product.images ?? [],
    tags: product.tags ?? [],
    createdAt: product.createdAt.toISOString(),
  };
}

router.get("/products", async (req, res): Promise<void> => {
  const parsed = ListProductsQueryParams.safeParse(req.query);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const { search, categoryId, supplierId, minPrice, maxPrice, page = 1, limit = 20 } = parsed.data;

  const inStock = req.query.inStock === "true" ? true : req.query.inStock === "false" ? false : null;
  const verifiedOnly = req.query.verifiedOnly === "true";
  const minRating = req.query.minRating != null ? Number(req.query.minRating) : null;
  const maxMoq = req.query.maxMoq != null ? Number(req.query.maxMoq) : null;
  const sort = typeof req.query.sort === "string" ? req.query.sort : "newest";

  const where: Prisma.ProductWhereInput = {};

  if (search) where.name = { contains: search, mode: "insensitive" };
  if (categoryId != null) where.categoryId = categoryId;
  if (supplierId != null) where.supplierId = supplierId;
  if (minPrice != null || maxPrice != null) {
    where.minPrice = {
      ...(minPrice != null ? { gte: minPrice } : {}),
      ...(maxPrice != null ? { lte: maxPrice } : {}),
    };
  }
  if (inStock != null) where.inStock = inStock;
  if (maxMoq != null && !Number.isNaN(maxMoq)) where.minOrder = { lte: maxMoq };
  if (minRating != null && !Number.isNaN(minRating)) where.rating = { gte: minRating };

  if (verifiedOnly) {
    const verified = await prisma.supplier.findMany({
      where: { verified: true },
      select: { id: true },
    });
    const ids = verified.map((s) => s.id);
    if (ids.length === 0) {
      res.json({ items: [], total: 0, page, limit });
      return;
    }
    where.supplierId = { in: ids };
  }

  let orderBy: Prisma.ProductOrderByWithRelationInput = { createdAt: "desc" };
  if (sort === "price_asc") orderBy = { minPrice: "asc" };
  else if (sort === "price_desc") orderBy = { minPrice: "desc" };
  else if (sort === "rating") orderBy = { rating: "desc" };
  else if (sort === "moq") orderBy = { minOrder: "asc" };

  const [total, items] = await Promise.all([
    prisma.product.count({ where }),
    prisma.product.findMany({
      where,
      orderBy,
      take: limit,
      skip: (page - 1) * limit,
    }),
  ]);

  const enriched = await Promise.all(items.map(enrichProduct));
  res.json({ items: enriched, total, page, limit });
});

router.get("/products/featured", async (_req, res): Promise<void> => {
  const items = await prisma.product.findMany({
    where: { featured: true },
    take: 12,
  });

  const enriched = await Promise.all(items.map(enrichProduct));
  res.json(GetFeaturedProductsResponse.parse(enriched));
});

router.get("/products/:id", async (req, res): Promise<void> => {
  const rawId = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = GetProductParams.safeParse({ id: parseInt(rawId, 10) });
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const product = await prisma.product.findUnique({ where: { id: params.data.id } });
  if (!product) {
    res.status(404).json({ error: "Product not found" });
    return;
  }

  const enriched = await enrichProduct(product);
  res.json(GetProductResponse.parse(enriched));
});

router.post("/products", requireClerkAuth, async (req, res): Promise<void> => {
  const parsed = CreateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const dbUser = await getAuthenticatedDbUser(req);
  if (!dbUser) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }
  if (!isSellerOrAdmin(dbUser)) {
    res.status(403).json({ error: "Forbidden — only sellers can manage products" });
    return;
  }
  if (!canAccessSupplier(dbUser, parsed.data.supplierId)) {
    res.status(403).json({
      error: "Forbidden — you can only create products for your linked supplier",
    });
    return;
  }

  const { getSupplierEntitlements } = await import("../lib/shop");
  const entitlements = await getSupplierEntitlements(parsed.data.supplierId);
  const productCount = await prisma.product.count({
    where: { supplierId: parsed.data.supplierId },
  });
  if (productCount >= entitlements.maxProducts) {
    res.status(402).json({
      error: `Your ${entitlements.planCode} plan allows ${entitlements.maxProducts} products. Upgrade your shop plan to list more.`,
      planCode: entitlements.planCode,
      maxProducts: entitlements.maxProducts,
      productCount,
    });
    return;
  }

  // Only admins may feature products
  const createData = { ...parsed.data };
  if (createData.featured && !isAdmin(dbUser)) {
    delete createData.featured;
  }

  const rating =
    "rating" in parsed.data && parsed.data.rating != null
      ? (parsed.data as { rating?: number }).rating
      : null;

  const product = await prisma.product.create({
    data: {
      ...createData,
      images: parsed.data.images ?? [],
      tags: parsed.data.tags ?? [],
      rating: rating ?? undefined,
    },
  });

  const enriched = await enrichProduct(product);
  res.status(201).json(GetProductResponse.parse(enriched));
});

router.patch("/products/:id", requireClerkAuth, async (req, res): Promise<void> => {
  const rawId = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = UpdateProductParams.safeParse({ id: parseInt(rawId, 10) });
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateProductBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const dbUser = await getAuthenticatedDbUser(req);
  if (!dbUser) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }

  const existing = await prisma.product.findUnique({ where: { id: params.data.id } });
  if (!existing) {
    res.status(404).json({ error: "Product not found" });
    return;
  }
  if (!canAccessSupplier(dbUser, existing.supplierId)) {
    res.status(403).json({ error: "Forbidden — you cannot edit this product" });
    return;
  }
  if (!isSellerOrAdmin(dbUser)) {
    res.status(403).json({ error: "Forbidden — only sellers can manage products" });
    return;
  }

  const data: Prisma.ProductUpdateInput = { ...parsed.data };
  if ("featured" in data && data.featured != null && !isAdmin(dbUser)) {
    delete data.featured;
  }
  const rating = (parsed.data as { rating?: number }).rating;
  if (rating != null) data.rating = rating;

  const product = await prisma.product.update({
    where: { id: params.data.id },
    data,
  });
  const enriched = await enrichProduct(product);
  res.json(UpdateProductResponse.parse(enriched));
});

router.delete("/products/:id", requireClerkAuth, async (req, res): Promise<void> => {
  const rawId = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = DeleteProductParams.safeParse({ id: parseInt(rawId, 10) });
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const dbUser = await getAuthenticatedDbUser(req);
  if (!dbUser) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }
  if (!isSellerOrAdmin(dbUser)) {
    res.status(403).json({ error: "Forbidden — only sellers can manage products" });
    return;
  }

  const existing = await prisma.product.findUnique({ where: { id: params.data.id } });
  if (!existing) {
    res.status(404).json({ error: "Product not found" });
    return;
  }
  if (!canAccessSupplier(dbUser, existing.supplierId)) {
    res.status(403).json({ error: "Forbidden — you cannot delete this product" });
    return;
  }

  await prisma.product.delete({ where: { id: params.data.id } });
  res.status(204).send();
});

export default router;