# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Karm Baba B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional search and filter
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a supplier or product
      parameters:
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /dashboard/supplier/{id}:
    get:
      operationId: getSupplierDashboard
      tags: [dashboard]
      summary: Get supplier-specific dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDashboard"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, supplierId, categoryId, minPrice, maxPrice, unit, minOrder, imageUrl, inStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        supplierVerified:
          type: ["boolean", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        minPrice:
          type: number
        maxPrice:
          type: number
        unit:
          type: string
        minOrder:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, supplierId, categoryId, minPrice, maxPrice, unit, minOrder, imageUrl]
      properties:
        name:
          type: string
        description:
          type: string
        supplierId:
          type: integer
        categoryId:
          type: integer
        minPrice:
          type: number
        maxPrice:
          type: number
        unit:
          type: string
        minOrder:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        minPrice:
          type: number
        maxPrice:
          type: number
        unit:
          type: string
        minOrder:
          type: integer
        imageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
    ProductListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Supplier:
      type: object
      required: [id, companyName, location, verified, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        country:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        yearsInBusiness:
          type: ["integer", "null"]
        employeeCount:
          type: ["string", "null"]
        mainProducts:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        responseRate:
          type: ["number", "null"]
        responseTime:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Rfq:
      type: object
      required: [id, productName, quantity, unit, buyerName, buyerEmail, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        buyerId:
          type: ["integer", "null"]
        buyerName:
          type: string
        buyerEmail:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        targetPrice:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, responded, accepted, rejected]
        createdAt:
          type: string
...
    Review:
      type: object
      required: [id, rating, comment, reviewerName, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        reviewerId:
          type: ["integer", "null"]
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating, comment, reviewerName]
      properties:
        supplierId:
          type: integer
        productId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
...
    SupplierDashboard:
      type: object
      required: [supplier, productCount, rfqCount, pendingRfqs, recentRfqs, totalViews]
      properties:
        supplier:
          $ref: "#/components/schemas/Supplier"
        productCount:
          type: integer
        rfqCount:
          type: integer
        pendingRfqs:
          type: integer
        recentRfqs:
          type: array
          items:
            $ref: "#/components/schemas/Rfq"
        totalViews:
          type: integer
          description: Reserved for profile-view analytics; currently always 0 (not tracked).
...